+++ recipes/package_repo.rb
@@ -19,11 +19,21 @@ when 'debian'
     components ['main']
     keyserver node['percona']['apt']['keyserver']
     key node['percona']['apt']['key']
-    not_if "apt-key list | grep #{node['percona']['apt']['key'][-8, 8]}"
+  end
+
+  if node['percona']['version'].to_f >= 8.0
+    node['percona']['repositories'].each do |repo|
+      apt_repository "percona-#{repo}" do
+        uri "http://repo.percona.com/#{repo}/apt"
+        components ['main']
+        keyserver node['percona']['apt']['keyserver']
+        key node['percona']['apt']['key']
+      end
+    end
   end
 
 when 'rhel'
-  yum_repository "percona-#{node['kernel']['machine']}" do
+  yum_repository 'percona' do
     description node['percona']['yum']['description']
     baseurl "#{node['percona']['yum']['baseurl']}/$basearch"
     gpgkey node['percona']['yum']['gpgkey']
@@ -31,8 +41,20 @@ when 'rhel'
     sslverify node['percona']['yum']['sslverify']
   end
 
+  if node['percona']['version'].to_f >= 8.0
+    node['percona']['repositories'].each do |repo|
+      yum_repository "percona-#{repo}" do
+        description node['percona']['yum']['description'] + ' - ' + repo
+        baseurl "http://repo.percona.com/#{repo}/yum/release/$releasever/RPMS/$basearch"
+        gpgkey node['percona']['yum']['gpgkey']
+        gpgcheck node['percona']['yum']['gpgcheck']
+        sslverify node['percona']['yum']['sslverify']
+      end
+    end
+  end
+
   yum_repository 'percona-noarch' do
-    description node['percona']['yum']['description']
+    description node['percona']['yum']['description'] + ' - noarch'
     baseurl "#{node['percona']['yum']['baseurl']}/noarch"
     gpgkey node['percona']['yum']['gpgkey']
     gpgcheck node['percona']['yum']['gpgcheck']