+++ recipes/server.rb
@@ -5,21 +5,14 @@
 
 include_recipe 'percona::package_repo'
 
-version = node['percona']['version']
-
 # install packages
 case node['platform_family']
 when 'debian'
-  node.default['percona']['server']['package'] = "percona-server-server-#{version}"
-
   package node['percona']['server']['package'] do
     options '--force-yes'
     action node['percona']['server']['package_action'].to_sym
   end
 when 'rhel'
-  node.default['percona']['server']['package'] = "Percona-Server-server-#{version.tr('.', '')}"
-  node.default['percona']['server']['shared_pkg'] = "Percona-Server-shared-#{version.tr('.', '')}"
-
   # we need mysqladmin
   include_recipe 'percona::client'