+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This file is used to list changes made in each version of the percona cookbook.
 - Add support for Ubuntu 20.04 for 5.7 only
 - Add support for CentOS 8
 - Add cluster suite and tests to test cluster recipe
+- Add support for Percona 8.0 and default to that version
 
 ### Fixed